+++ src/prompt/base.ts
@@ -29,7 +29,6 @@ export class Prompt<T> {
     retryFn?: () => Promise<T | undefined>,
   ) => void | Promise<void>;
   protected preNewLine: number;
-  protected onCtrlR?: () => boolean | void | Promise<boolean | void>;
   protected returnToMainMenu: Opts.ReturnToMainMenuOpt;
   protected returnToMainMenuLabel: string;
 
@@ -52,13 +51,8 @@ export class Prompt<T> {
     this.maxAttempts = opts.maxAttempts;
     this.onExceededAttempts = opts.onExceededAttempts ?? onExceededAttempts;
     this.preNewLine = opts.preNewLine ?? 1;
-    this.onCtrlR = opts.onCtrlR;
     this.returnToMainMenu = opts.returnToMainMenu ?? 'off';
     this.returnToMainMenuLabel = opts.returnToMainMenuLabel ?? 'Return to Main Menu';
-
-    if (this.onCtrlR && (this.prefix === '' || this.prefix === null)) {
-      this.prefix = ' ';
-    }
   }
 
   protected async start(): Promise<void> {